mina_signer::pubkey

Struct CompressedPubKey

pub struct CompressedPubKey {
    pub x: BaseField,
    pub is_odd: bool,
}

Compressed public keys consist of x-coordinate and y-coordinate parity.

Fields
x: BaseField

X-coordinate

is_odd: bool

Parity of y-coordinate

Implementations

impl CompressedPubKey

pub fn into_address(&self) -> String

Serialize compressed public key into corresponding Mina address

pub fn from_bytes(bytes: &[u8]) -> Result<Self>

Deserialize compressed public key from bytes

Errors

Will give error if bytes do not match certain requirements.

pub fn from_hex(public_hex: &str) -> Result<Self>

Deserialize compressed public key from hex

Errors

Will give error if hex string does not match certain requirements.

pub fn from_secret_key(sec_key: SecKey) -> Self

Create compressed public key from a secret key

pub fn from_address(address: &str) -> Result<Self>

Deserialize Mina address into compressed public key (via an uncompressed PubKey)

Errors

Will give error if PubKey::from_address() returns error.

pub fn empty() -> Self

The empty CompressedPubKey value that is used as public_key in empty account and None value for calculating the hash of Option<CompressedPubKey>, etc.

pub fn to_bytes(&self) -> Vec<u8>

Deserialize compressed public key into bytes

pub fn to_hex(&self) -> String

Deserialize compressed public key into hex