Trait mina_signer::Signer

pub trait Signer {
    fn sign<S: Signable>(&mut self, kp: Keypair, input: S) -> Signature;

    fn verify<S: Signable>(
        &mut self, 
        sig: Signature, 
        pub_key: PubKey, 
        input: S
    ) -> bool;
}

Signer interface for signing Signable inputs and verifying Signatures using Keypairs and PubKeys

Required methods

fn sign<S: Signable>(&mut self, kp: Keypair, input: S) -> Signature

Sign input (see Signable) using keypair kp and return the corresponding signature.

fn verify<S: Signable>(
&mut self,
sig: Signature,
pub_key: PubKey,
input: S
) -> bool

Verify that the signature sig on input (see Signable) is signed with the secret key corresponding to pub_key. Return true if the signature is valid and false otherwise.